+++ src/_pytest/_code/__init__.py
@@ -1,10 +1,22 @@
-""" python inspection/code generation API """
-from .code import Code  # noqa
-from .code import ExceptionInfo  # noqa
-from .code import filter_traceback  # noqa
-from .code import Frame  # noqa
-from .code import getrawcode  # noqa
-from .code import Traceback  # noqa
-from .source import compile_ as compile  # noqa
-from .source import getfslineno  # noqa
-from .source import Source  # noqa
+"""Python inspection/code generation API."""
+from .code import Code
+from .code import ExceptionInfo
+from .code import filter_traceback
+from .code import Frame
+from .code import getfslineno
+from .code import getrawcode
+from .code import Traceback
+from .source import compile_ as compile
+from .source import Source
+
+__all__ = [
+    "Code",
+    "ExceptionInfo",
+    "filter_traceback",
+    "Frame",
+    "getfslineno",
+    "getrawcode",
+    "Traceback",
+    "compile",
+    "Source",
+]
