+++ src/_pytest/nodes.py
@@ -12,10 +12,10 @@ from typing import Union
 import py
 
 import _pytest._code
+from _pytest._code import getfslineno
 from _pytest._code.code import ExceptionChainRepr
 from _pytest._code.code import ExceptionInfo
 from _pytest._code.code import ReprExceptionInfo
-from _pytest._code.source import getfslineno
 from _pytest.compat import cached_property
 from _pytest.compat import TYPE_CHECKING
 from _pytest.config import Config
